docker/newsFeeder/requirements.txt
feedparser==6.0.10
beautifulsoup4==4.9.3
urllib3==1.26.6
dateparser==1.1.1
requests==2.25.1
newspaper3k==0.2.8
google-cloud-storage==2.1.0
python-dotenv==0.19.0
langchain==0.0.146  # Make sure to use a version that includes document_loaders, text_splitter, embeddings, and vectorstores